+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVQT9DDSZQ65610R6F56HWRF"
+        "01KVSJ3CN0SZSX84WGTXFPEW08"
     )
     r"""The unique identifier of the data source"""