+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVQT9DDZBSB2AGKHAGF5T495"
+        "01KVSJ3CN5GVNFTNG59G8V127V"
     )
     r"""The unique identifier of the data source"""